Refactor the textfield to match with structure Rename the textfield file name
